tests/wpt/mozilla/tests/mozilla/mozbrowser/redirect.html

<!doctype html>
<meta charset="utf-8">
<title>mozbrowserlocationchange with final url after redirect</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body></body>
<script>
    async_test(function(t) {
        var iframe = document.createElement("iframe");
        iframe.mozbrowser = "true";
        iframe.src = "redirect_init.html?pipe=status(302)|header(Location,redirect_final.html)";
        var i = 0;
        iframe.addEventListener("mozbrowserlocationchange", t.step_func(e => {
            switch (++i) {
                case 1:
                    assert_equals(e.detail.url, "about:blank");
                    break;
                case 2:
                    assert_equals(e.detail.url, new URL("redirect_final.html", location).href);
                    t.done();
                    break;
            }
        }));
        document.body.appendChild(iframe);
    });
</script>
